fixes to the demo script   - create current, last and today links   - respawn itself (allows runtime modifications of the script)   - more/cleaner debug information into log file   - cleanup of html representation (although that can still be improved...)
The demo script should now run continously on cyder.  The frequency of demos can be changed by editing demo.delay, which defines a sleep time in seconds between demo runs.  The cyder deployment lives in ~amerzky/mandelbrot/demo.  web page is at:

  http://cyder.cct.lsu.edu/saga-interop/mandelbrot/demo/
  http://cyder.cct.lsu.edu/saga-interop/mandelbrot/demo/today/last

A


git-svn-id: file://localhost/tmp/svn2git/svn@3345 defb5e50-622e-49ec-a68e-d72c7db87b45

+++ applications/mandelbrot/demo/run_demo.sh
@@ -12,35 +12,10 @@ if [ -z "$SAGA_LOCATION" ]; then
   die "SAGA_LOCATION needs to be set"
 fi
 
-export SAGA_DEMO_ROOT=$HOME/saga/mandelbrot/demo
-export SAGA_DEMO_LOG=$SAGA_DEMO_ROOT/demo.log
-export SAGA_DEMO_LOCK=$SAGA_DEMO_ROOT/demo.lock
-export SAGA_DEMO_STOP=$SAGA_DEMO_ROOT/demo.stop
-export SAGA_DEMO_DELAY=$SAGA_DEMO_ROOT/demo.delay
-
-export SAGA_MANDELBROT_INI=$SAGA_DEMO_ROOT/demo.ini
-
-# have only one demo script running at any time
-if [ -f $SAGA_DEMO_LOCK ]; then
-  echo ""
-  echo " Found lock file at $SAGA_DEMO_LOCK - demo is already running" 
-  echo ""
-  exit -1
-else
-  touch $SAGA_DEMO_LOCK
-
-  # stop signal MUST be stale here
-  rm -f $SAGA_DEMO_STOP 
-fi
-
-preprev="-"
-prev="-"
-this="-"
-
+function main
 {
   # run as long as we don't get signalled to stop
-  while [ ! -f $SAGA_DEMO_STOP ]
-  do
+  if [ ! -f $SAGA_DEMO_STOP ]; then
 
     export SAGA_DEMO_TIME=`/bin/date "+%H:%M"`
     export SAGA_DEMO_DATE=`/bin/date "+%Y-%m-%d"`
@@ -51,9 +26,9 @@ this="-"
 
     export SAGA_MANDELBROT_PWD=$SAGA_DEMO_HOME
 
-    preprev=$prev
-    prev=$this
-    this=$SAGA_DEMO_TIME
+    export SAGA_DEMO_PREPREV=$SAGA_DEMO_PREV
+    export SAGA_DEMO_PREV=$SAGA_DEMO_THIS
+    export SAGA_DEMO_THIS=$SAGA_DEMO_TIME
 
     echo "==============================="
     echo "running demo in $SAGA_DEMO_HOME"
@@ -62,9 +37,19 @@ this="-"
     mkdir -p $SAGA_DEMO_HOME || die "cannot create demo home $SAGA_DEMO_HOME"
     cd       $SAGA_DEMO_HOME || die "cannot cd to  demo home $SAGA_DEMO_HOME"
 
+    # prepare today link and set static links for current and last demo
+    rm -f $SAGA_DEMO_ROOT/today
+    ln -s $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/ $SAGA_DEMO_ROOT/today
+    touch $SAGA_DEMO_ROOT/today/last
+    rm -f $SAGA_DEMO_ROOT/today/last
+    touch $SAGA_DEMO_ROOT/today/current
+    mv -f $SAGA_DEMO_ROOT/today/current $SAGA_DEMO_ROOT/today/last
+    ln -s $SAGA_DEMO_HOME               $SAGA_DEMO_ROOT/today/current
+
     # re-create html for last demo, to fix 'next'
-    if [ -d $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$prev ]; then
-      $SAGA_DEMO_ROOT/demo2html.pl $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$prev $preprev $this
+    if [ -d $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$SAGA_DEMO_PREV ]; then
+      echo "re-creating html: demo2html.pl $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$SAGA_DEMO_PREV $SAGA_DEMO_PREPREV $SAGA_DEMO_THIS"
+      $SAGA_DEMO_ROOT/demo2html.pl $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$SAGA_DEMO_PREV $SAGA_DEMO_PREPREV $SAGA_DEMO_THIS
     fi
 
     # this line expects mandelbrot_master to be installed in $SAGA_LOCATION/bin,
@@ -77,7 +62,8 @@ this="-"
     echo "===============================" 
 
     # create html for the current demo - no 'next' nown yet
-    $SAGA_DEMO_ROOT/demo2html.pl $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$this $prev -
+    echo "creating html: demo2html.pl $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$SAGA_DEMO_THIS $SAGA_DEMO_PREV -"
+    $SAGA_DEMO_ROOT/demo2html.pl $SAGA_DEMO_ROOT/demo-$SAGA_DEMO_DATE/$SAGA_DEMO_THIS $SAGA_DEMO_PREV -
 
     # new demo starts shortly after the last one finished (1 min sleep)
     if [ -f $SAGA_DEMO_DELAY ]; then
@@ -86,21 +72,59 @@ this="-"
       /bin/sleep 60
     fi
 
-  done
+    # relaunch ourself
+    $SAGA_DEMO_ROOT/run_demo.sh
 
-  rm -f $SAGA_DEMO_STOP
-  rm -f $SAGA_DEMO_LOCK
+  else # SAGA_DEMO_STOP exists
 
-  echo "===============================" 
-  echo "demo stopped "                   
-  echo "===============================" 
+    rm -f $SAGA_DEMO_STOP
+    rm -f $SAGA_DEMO_LOCK
 
+    echo "===============================" 
+    echo "demo stopped "                   
+    echo "===============================" 
 
-} 2>&1 > $SAGA_DEMO_LOG &
+  fi
+}
 
-echo ""
-echo " $0 is going into background mode."
-echo " Output will be captured in '$SAGA_DEMO_LOG'."
-echo " The script's pid is '$!'"
-echo " To cancel the demo, run 'stop_demo.sh'"
-echo ""
+export SAGA_DEMO_ROOT=$HOME/www/mandelbrot/demo
+export SAGA_DEMO_LOG=$SAGA_DEMO_ROOT/demo.log
+export SAGA_DEMO_LOCK=$SAGA_DEMO_ROOT/demo.lock
+export SAGA_DEMO_STOP=$SAGA_DEMO_ROOT/demo.stop
+export SAGA_DEMO_DELAY=$SAGA_DEMO_ROOT/demo.delay
+
+export SAGA_MANDELBROT_INI=$SAGA_DEMO_ROOT/demo.ini
+
+# the first run needs some 4extra setup
+if test "$SAGA_DEMO_SCRIPT_ITERATION" != "yes"; then
+
+  # have only one demo script running at any time
+  if [ -f $SAGA_DEMO_LOCK ]; then
+    echo ""
+    echo " Found lock file at $SAGA_DEMO_LOCK - demo is already running" 
+    echo ""
+    exit -1
+  else
+    touch $SAGA_DEMO_LOCK
+  
+    # stop signal MUST be stale here
+    rm -f $SAGA_DEMO_STOP 
+  fi
+
+  export SAGA_DEMO_PREPREV="-"
+  export SAGA_DEMO_PREV="-"
+  export SAGA_DEMO_THIS="-"
+
+  echo ""
+  echo " $0 is going into background mode."
+  echo " Output will be captured in '$SAGA_DEMO_LOG'."
+  echo " The script's pid is '$!'"
+  echo " To cancel the demo, run 'stop_demo.sh'"
+  echo ""
+
+  # don't run that init phase again
+  export SAGA_DEMO_SCRIPT_ITERATION=yes
+fi
+
+main 2>&1 >> $SAGA_DEMO_LOG &
+
